refactor: Consolidate PCA mathematical validation tests into main test file
- Move 11 mathematical validation tests from test_pca_validation.py to test_pca.py
- Delete separate validation file to maintain cleaner repo structure
- Add TestPCAMathematicalValidation class with comprehensive tests:
  * Shape/type validation and orthonormal loadings check
  * Trace accounting (sum of per-feature variance = sum of eigenvalues)
  * Total fraction explained bounds [0,1] and equals 1 with all PCs
  * Per-feature fraction bounds validation
  * Standardization produces population variance of 1
  * ddof consistency between PCA and variance calculations
  * Explicit loadings orthonormality test
  * Feature metrics DataFrame alignment and consistency
  * Component selection threshold behavior
  * Mahalanobis distance validation in 1D and nD
  * Reconstruction error monotonicity with more components

Add 3 specialized fixtures for mathematical validation:
- controlled_spectrum_data: Low-rank structure with known spectrum
- diagonal_covariance_data: Independent features with known eigenvalues
- correlated_pairs_data: Pairs of correlated features

All 88 PCA tests passing (77 original + 11 new), ensuring mathematical correctness.
